+++ packages/jupyter-chat/src/index.ts
@@ -13,5 +13,3 @@ export * from './registers';
 export * from './selection-watcher';
 export * from './types';
 export * from './widgets';
-export * from './widgets/multiChatPanel';
-export * from './token';
