+++ packages/jupyter-chat/src/widgets/index.ts
@@ -6,4 +6,4 @@
 export * from './chat-error';
 export * from './chat-sidebar';
 export * from './chat-widget';
-export * from './multiChatPanel';
+export * from './multichat-panel';
